refactor(ui->storage->folders): use dynamic routes
